fix style to use spacing, split component for inputs
